Create fann_aFirst.py
I don't get git to push only what it should, so now this way. This file uses another file that you can create yourself by un-commenting a line in the code here, or from the other file (will be put up on github next). If you store that second one in a folder "project_data" everything should work fine.

That's a xth version of a nn with fann; doesn't work well yet - 10.000 of 28.000 were miss-classified (in the training set).